requirements.txt
numpy>=1.24
PyYAML>=6.0
qiskit>=2.0
qiskit-aer>=0.14
scipy>=1.10
